revert back to just one build script, features will load dynamically
