+++ package.json
@@ -34,7 +34,6 @@
     "grunt-string-replace": "~1.2.0",
     "grunt-targethtml": "~0.2.6",
     "less": "~1.7.5",
-    "lodash": "^4.17.4",
     "rjs-build-analysis": "~0.0.3"
   },
   "license": [
