--- src/scripts/config.compiled.js
@@ -1,22 +1,0 @@
-(function () {
-  'use strict';
-
-  define(function () {
-    return function(settings){
-      console.info(settings.environment, settings.conceptCoach.base);
-      require.config({
-        paths: {
-          aloha: 'aloha',
-          OpenStaxConceptCoach: settings.conceptCoach.base + 'full-build.min'
-        },
-        shim: {
-          OpenStaxConceptCoach: {
-            deps: ['css!' + settings.conceptCoach.base + 'main.min'],
-            exports: 'OpenStaxConceptCoach'
-          }
-        }
-      });
-    }
-  });
-
-})();
