+++ src/scripts/helpers/backbone/views/content.coffee
@@ -1,0 +1,5 @@
+define (require) ->
+  feature = require('cs!helpers/feature')
+  editable = require('cs!helpers/backbone/views/editable')
+  mediaComponent = require('cs!helpers/backbone/views/media-component')
+  return feature('authoring', editable, mediaComponent)
