+++ src/scripts/helpers/underscore.deepExtend.coffee
@@ -1,0 +1,6 @@
+define (require) ->
+  _ = require('underscore')
+  underscoreDeepExtend = require('underscoreDeepExtend')
+  _.mixin(deepExtend: underscoreDeepExtend(_))
+
+  return _
