+++ src/scripts/main.js
@@ -3,8 +3,9 @@
 
   // Load the config
   require(['config', 'settings'], function(config, settings){
-    require(['config.' + settings.environment, 'cs!loader'], function(configureFor, loader){
-      configureFor(settings);
+    require(['settings/config', 'cs!loader', 'cs!helpers/underscore.deepExtend'], function(configureFor, loader, _){
+      // some paths and shims need to be configured based on settings.
+      configureFor(settings, _);
       // Load the application after the config
       loader.init();
     });
