+++ src/scripts/session-stub.coffee
@@ -1,0 +1,5 @@
+define (require) ->
+  stubModel =
+    update: -> null
+    get: -> null
+    off: -> null
